backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
qdrant-client>=1.12.0
python-multipart>=0.0.9
langgraph>=0.2.0
langchain>=0.3.0
langsmith>=0.2.0
langchain-community>=0.3.0
langchain-classic>=1.0.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-cohere>=0.4.0
pypdf>=5.0.0
ragas>=0.2.0
pandas>=2.0.0
nltk>=3.8.0
rapidfuzz>=3.0.0
